i'm in love with a british film star

+++ code/game/objects/items.dm
@@ -95,14 +95,15 @@
 		src:holder = null
 	for(var/x in actions)
 		qdel(x)
-		x = null
 	/*  BROKEN, FUCK BYOND
 	if(hasvar(src, "my_atom"))
 		src:my_atom = null*/
-	qdel(daemon_datum)
-	qdel(hidden_uplink)
-	hidden_uplink = null
-	daemon_datum = null
+	if (daemon_datum)
+		qdel(daemon_datum)
+		daemon_datum = null
+	if (hidden_uplink)
+		qdel(hidden_uplink)
+		hidden_uplink = null
 	virus2.Cut()
 	..()
